created react function component for nav and pass in an array
