+++ src/assignments/project-1/components/footer.js
@@ -1,0 +1,8 @@
+
+import React from 'react'
+const Footer = () => (
+	<footer className="main-footer">Images are copyrighted by the original owner. </footer>
+)
+
+export default Footer
+
