+++ src/assignments/project-1/components/header.js
@@ -1,0 +1,6 @@
+import React from 'react' 
+const Header = ({ imgSrc, text, children }) => ( 
+	<header className="main-head"><img className="header-photo" src={imgSrc} />{text}{children}</header>  
+) 
+
+export default Header
