+++ src/assignments/project-1/components/img.js
@@ -1,25 +1,12 @@
-import React from 'react' 
+import React from 'react'
 
- const Img = ({ className, small, large, alt, style}) => (
+const Img = ({ className, small, large, alt, style }) => (
 	<picture className={className}>
 		<source media="(max-width: 767px)" srcSet={small} />
 		<source media="(min-width: 768px)" srcSet={large} />
-		<img src={large} alt={alt} style={style} /> 
+		<img src={large} alt={alt} style={style} />
 	</picture>
 );
-/*class Img extends React.Component {
-
-
-render() {
-    const {upload, ...inputProps} = this.props
-    return <input {...inputProps} multiple type="file" onChange={this.handleFiles} />
-}
-
-}
-
-Uploader.propTypes = {
-upload: PropTypes.func.isRequired,
-}
-*/
+ 
 
 export default Img
