+++ src/assignments/project-1/components/list.js
@@ -1,0 +1,8 @@
+import React from 'react'
+function List({ items }) {
+    return items.map(item => (
+        <li key={item.name} className="main-li">{item.name}</li>
+    ));
+}
+
+export default List
