+++ src/assignments/project-1/components/nav.js
@@ -1,0 +1,16 @@
+import React from 'react'
+import List from './list'
+
+
+
+function Nav({ listItems }) {
+    return (
+        <aside className="main-nav">
+            <ul>
+                <List items={listItems} />
+            </ul>
+        </aside>
+    );
+}
+ 
+export default Nav
